LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/valid-content-type.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.
CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'assert'. Expected a ';' following a targeted import declaration.

FAIL text/json assert_unreached: [object Event] Reached unreachable code
FAIL application/json assert_unreached: [object Event] Reached unreachable code
FAIL text/html+json assert_unreached: [object Event] Reached unreachable code
FAIL image/svg+json assert_unreached: [object Event] Reached unreachable code
FAIL text/json;boundary=something assert_unreached: [object Event] Reached unreachable code
FAIL text/json;foo=bar assert_unreached: [object Event] Reached unreachable code
